Refactor: 타입따로 분리 작업
