+++ src/types/auth.ts
@@ -1,0 +1,9 @@
+export interface AuthRequest {
+  email: string;
+  password: string;
+}
+
+export interface SignupRequest extends AuthRequest {
+  username: string;
+  userType?: string;
+}
